crillerium/requirements.txt
pandas
numpy
scikit-learn
lightgbm
matplotlib
seaborn
xlrd
openpyxl
joblib
python-docx
scipy
